Split humanizers into separate files
