+++ gluco-check-core/src/main/i18n/humanizers/index.ts
@@ -1,0 +1,18 @@
+import bloodSugar from './BloodSugar';
+import cannulaAge from './CannulaAge';
+import sensorAge from './SensorAge';
+import pumpBattery from './PumpBattery';
+import carbsOnBoard from './CarbsOnBoard';
+import insulinOnBoard from './InsulinOnBoard';
+
+/**
+ * Humanizers turn an internal concept such as a DiabetesPointer, into human text
+ */
+export default {
+  bloodSugar,
+  cannulaAge,
+  sensorAge,
+  pumpBattery,
+  carbsOnBoard,
+  insulinOnBoard,
+};
