+++ gluco-check-core/tsconfig.json
@@ -1,6 +1,9 @@
 {
   "extends": "./node_modules/gts/tsconfig-google.json",
   "compilerOptions": {
+    // FIXME: forceConsistentCasingInFileNames this may be removed again if this gets fixed:
+    // https://github.com/microsoft/TypeScript/issues/25460 
+    "forceConsistentCasingInFileNames": false, 
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
     "rootDir": ".",
